[CP-3674][App Refactor][Harmony] Sounds - cleanup (#2681)
