+++ libs/app-localize/utils/src/lib/locales/en-US.json
@@ -93,19 +93,21 @@
   "general.welcomeScreen.subtitle": "Connect your device to the computer.",
   "general.welcomeScreen.title": "Welcome to Mudita Center",
   "general.welcomeScreen.troubleshootingButton.text": "My device doesn’t show up after connecting",
+  "harmony.manageFiles.alarmFiles.emptyStateDescription": "Add alarm files from your computer and they’ll transfer to your device automatically.",
+  "harmony.manageFiles.confirmDelete.modal.cancelButtonText": "Cancel",
+  "harmony.manageFiles.confirmDelete.modal.confirmButtonText": "Delete {fileCount, plural, one {file} other {files}}",
   "harmony.manageFiles.confirmDelete.modal.description": "This will remove the {fileCount, plural, one {file} other {files}} from Harmony.",
-  "harmony.manageFiles.confirmDelete.modal.primaryButtonText": "Delete {fileCount, plural, one {file} other {files}}",
-  "harmony.manageFiles.confirmDelete.modal.secondaryButtonText": "Cancel",
   "harmony.manageFiles.confirmDelete.modal.title": "Delete {fileCount, plural, one {file} other {# files}}?",
   "harmony.manageFiles.deleteFailed.all.modal.description": "{failedCount, plural, one {The file is still on Harmony, please try again.} other {The files are still on Harmony, please try again.}}",
   "harmony.manageFiles.deleteFailed.all.modal.title": "{failedCount, plural, one {Deleting failed} other {Deleting failed for # files}}",
-  "harmony.manageFiles.deleteFailed.modal.secondaryButtonText": "Cancel",
+  "harmony.manageFiles.deleteFailed.modal.closeButtonText": "Cancel",
   "harmony.manageFiles.deleteFailed.some.modal.description": "{succeededCount, plural, =0 {No files} one {# file} other {# files}} successfully deleted, {failedCount, plural, one {# file} other {# files}} not deleted:",
   "harmony.manageFiles.deleteFailed.some.modal.title": "Deleting some files failed",
   "harmony.manageFiles.deleting.modal.title": "Deleting, please wait...",
   "harmony.manageFiles.exportFailed.errorLabels.export.unknown": "export failed",
   "harmony.manageFiles.otherFilesOtherLabelText": "Other",
   "harmony.manageFiles.otherFilesSystemLabelText": "System",
+  "harmony.manageFiles.relaxationFiles.emptyStateDescription": "Add relaxation files from your computer and they’ll transfer to your device automatically.",
   "harmony.manageFiles.summary.header": "Harmony Sounds",
   "harmony.manageFiles.upload.validationFailure.duplicatesDescription": "{filesCount, plural, one {A file with this name already exists} other {Files with these names already exist}}. Please rename the duplicate {filesCount, plural, one {file} other {files}} and try again.",
   "harmony.manageFiles.upload.validationFailure.fileTooLargeDescription": "Files must be less than 2GB.\nSelect a smaller file and try again.",
