+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-delete-flow/manage-files-delete-flow.tsx
@@ -5,12 +5,12 @@
 
 import { FunctionComponent, useCallback, useEffect, useState } from "react"
 import { delayUntil } from "app-utils/common"
-import { FileManagerFile } from "./manage-files.types"
-import { ManageFilesDeletingModal } from "./manage-files-deleting-modal"
+import { FileManagerFile } from "../manage-files.types"
+import { ManageFilesDeleteProgressModal } from "./manage-files-delete-progress-modal"
 import {
-  ManageFilesConfirmDeleteModal,
   ManageFilesConfirmDeleteModalProps,
-} from "./manage-files-confirm-delete-modal"
+  ManageFilesDeleteConfirmModal,
+} from "./manage-files-delete-confirm-modal"
 import {
   ManageFilesDeleteFailedModal,
   ManageFilesDeleteFailedModalProps,
@@ -26,7 +26,7 @@ export interface ManageFilesDeleteFlowProps {
   opened: boolean
   selectedFiles: FileManagerFile[]
   onClose: VoidFunction
-  handleDeleteFile: (fileId: string) => Promise<void>
+  deleteFile: (fileId: string) => Promise<void>
   onDeleteSuccess?: () => Promise<void>
   onPartialDeleteFailure?: (failedFiles: FileManagerFile[]) => Promise<void>
   deleteFlowMessages: ManageFilesDeleteFlowMessages
@@ -38,7 +38,7 @@ export const ManageFilesDeleteFlow: FunctionComponent<
   opened,
   onClose,
   selectedFiles,
-  handleDeleteFile,
+  deleteFile,
   onDeleteSuccess,
   onPartialDeleteFailure,
   deleteFlowMessages,
@@ -52,40 +52,40 @@ export const ManageFilesDeleteFlow: FunctionComponent<
     setFlowState(opened ? ManageFilesDeleteFlowState.ConfirmDelete : null)
   }, [opened])
 
-  const handleDeleteFiles = useDeleteFilesHandler({
+  const deleteFiles = useDeleteFilesHandler({
     selectedFiles,
-    handleDeleteFile,
+    deleteFile,
     onDeleteSuccess,
     onSetFlowState: setFlowState,
     onSetFailedFiles: setFailedFiles,
   })
 
-  const handleConfirmDeleteClick = useCallback(() => {
+  const confirmDelete = useCallback(() => {
     setFlowState(ManageFilesDeleteFlowState.Deleting)
-    void delayUntil(handleDeleteFiles(), 500)
-  }, [handleDeleteFiles])
+    void delayUntil(deleteFiles(), 250)
+  }, [deleteFiles])
 
-  const handleDeleteFailedClose = useCallback(async () => {
+  const closeDeleteFailedModal = useCallback(async () => {
     onPartialDeleteFailure && (await onPartialDeleteFailure(failedFiles))
     setFlowState(null)
   }, [failedFiles, onPartialDeleteFailure])
 
   return (
     <>
-      <ManageFilesConfirmDeleteModal
+      <ManageFilesDeleteConfirmModal
         opened={flowState === ManageFilesDeleteFlowState.ConfirmDelete}
         onClose={onClose}
-        onPrimaryButtonClick={handleConfirmDeleteClick}
-        onSecondaryButtonClick={onClose}
+        onConfirm={confirmDelete}
+        onCancel={onClose}
         fileCount={selectedFiles.length}
         messages={deleteFlowMessages}
       />
-      <ManageFilesDeletingModal
+      <ManageFilesDeleteProgressModal
         opened={flowState === ManageFilesDeleteFlowState.Deleting}
       />
       <ManageFilesDeleteFailedModal
         opened={flowState === ManageFilesDeleteFlowState.DeleteFailed}
-        onClose={handleDeleteFailedClose}
+        onClose={closeDeleteFailedModal}
         messages={deleteFlowMessages}
         failedFiles={failedFiles}
         allFiles={selectedFiles}
