+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-delete-flow/manage-files-delete-progress-modal.tsx
@@ -18,7 +18,7 @@ interface Props {
   opened: boolean
 }
 
-export const ManageFilesDeletingModal: FunctionComponent<Props> = ({
+export const ManageFilesDeleteProgressModal: FunctionComponent<Props> = ({
   opened,
 }) => {
   return (
