+++ libs/devices/harmony/routes/src/lib/screens/manage-files/harmony-manage-files.messages.tsx
@@ -1,0 +1,120 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import { defineMessages } from "app-localize/utils"
+
+export const HarmonyManageFilesMessages = defineMessages({
+  pageTitle: {
+    id: "page.manageFiles.title",
+  },
+  summaryHeader: {
+    id: "harmony.manageFiles.summary.header",
+  },
+  confirmDeleteModalTitle: {
+    id: "harmony.manageFiles.confirmDelete.modal.title",
+  },
+  confirmDeleteModalDescription: {
+    id: "harmony.manageFiles.confirmDelete.modal.description",
+  },
+  confirmDeleteModalConfirmButtonText: {
+    id: "harmony.manageFiles.confirmDelete.modal.confirmButtonText",
+  },
+  confirmDeleteModalCancelButtonText: {
+    id: "harmony.manageFiles.confirmDelete.modal.cancelButtonText",
+  },
+  otherFilesSystemLabelText: {
+    id: "harmony.manageFiles.otherFilesSystemLabelText",
+  },
+  otherFilesOtherLabelText: {
+    id: "harmony.manageFiles.otherFilesOtherLabelText",
+  },
+  deleteFailedAllModalTitle: {
+    id: "harmony.manageFiles.deleteFailed.all.modal.title",
+  },
+  deleteFailedSomeModalTitle: {
+    id: "harmony.manageFiles.deleteFailed.some.modal.title",
+  },
+  deleteFailedAllModalDescription: {
+    id: "harmony.manageFiles.deleteFailed.all.modal.description",
+  },
+  deleteFailedDescriptionModalDescription: {
+    id: "harmony.manageFiles.deleteFailed.some.modal.description",
+  },
+  deleteFailedModalCloseButtonText: {
+    id: "harmony.manageFiles.deleteFailed.modal.closeButtonText",
+  },
+  uploadValidationFailureModalTitle: {
+    id: "harmony.manageFiles.upload.validationFailure.modalTitle",
+  },
+  uploadValidationFailureDuplicatesDescription: {
+    id: "harmony.manageFiles.upload.validationFailure.duplicatesDescription",
+  },
+  uploadValidationFailureInsufficientMemoryDescription: {
+    id: "harmony.manageFiles.upload.validationFailure.insufficientMemoryDescription",
+  },
+  uploadValidationFailureFileTooLargeDescription: {
+    id: "harmony.manageFiles.upload.validationFailure.fileTooLargeDescription",
+  },
+  uploadValidationFailureModalCloseButtonText: {
+    id: "harmony.manageFiles.deleteFailed.modal.closeButtonText",
+  },
+  uploadFailedAllModalTitle: {
+    id: "harmony.manageFiles.uploadFailed.all.modal.title",
+  },
+  uploadFailedSomeModalTitle: {
+    id: "harmony.manageFiles.uploadFailed.some.modal.title",
+  },
+  uploadFailedAllModalDescription: {
+    id: "harmony.manageFiles.uploadFailed.all.modal.description",
+  },
+  uploadFailedSomeModalDescription: {
+    id: "harmony.manageFiles.uploadFailed.some.modal.description",
+  },
+  uploadFailedModalCloseButtonText: {
+    id: "harmony.manageFiles.uploadFailed.modal.closeButtonText",
+  },
+  uploadFailedAllUnknownError: {
+    id: "harmony.manageFiles.uploadFailed.all.unknownError",
+  },
+  uploadFailedAllDuplicatesError: {
+    id: "harmony.manageFiles.uploadFailed.all.duplicatesError",
+  },
+  uploadFailedAllNotEnoughMemoryError: {
+    id: "harmony.manageFiles.uploadFailed.all.notEnoughMemoryError",
+  },
+  uploadFailedAllFileTooLargeError: {
+    id: "harmony.manageFiles.uploadFailed.all.fileTooLargeError",
+  },
+  uploadFailedErrorLabelUploadUnknown: {
+    id: "harmony.manageFiles.uploadFailed.errorLabels.upload.unknown",
+  },
+  exportFailedErrorLabelExportUnknown: {
+    id: "harmony.manageFiles.exportFailed.errorLabels.export.unknown",
+  },
+  uploadFailedErrorLabelDuplicate: {
+    id: "harmony.manageFiles.uploadFailed.errorLabels.duplicate",
+  },
+  uploadFailedErrorLabelCancelled: {
+    id: "harmony.manageFiles.uploadFailed.errorLabels.cancelled",
+  },
+  uploadFailedErrorLabelTooBig: {
+    id: "harmony.manageFiles.uploadFailed.errorLabels.tooBig",
+  },
+  uploadFailedErrorLabelFileTooLarge: {
+    id: "harmony.manageFiles.uploadFailed.errorLabels.fileTooLarge",
+  },
+  uploadingModalTitle: {
+    id: "harmony.manageFiles.uploading.modal.title",
+  },
+  uploadingModalCloseButtonText: {
+    id: "harmony.manageFiles.uploading.modal.closeButtonText",
+  },
+  relaxationFilesEmptyStateDescription: {
+    id: "harmony.manageFiles.relaxationFiles.emptyStateDescription",
+  },
+  alarmFilesEmptyStateDescription: {
+    id: "harmony.manageFiles.alarmFiles.emptyStateDescription",
+  },
+})
