--- libs/devices/harmony/routes/src/lib/screens/manage-files/map-to-harmony.utils.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) Mudita sp. z o.o. All rights reserved.
- * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
- */
-
-import { HarmonyFile } from "devices/harmony/models"
-
-export const sumFileSizes = (files: HarmonyFile[]) =>
-  files.reduce((acc, f) => acc + f.fileSize, 0)
